requirements.txt
torch
torchvision
timm
numpy
pandas
scikit-learn
Pillow
tqdm
safetensors
pyyaml
openpyxl
joblib
